Extract outcome comment builder to pr-fix-report
Move the `/fix` outcome message composition into a dedicated module under scripts/gh/pr-fix-report (index.mjs) and add unit tests (index.test.mjs). Rename the CLI file to scripts/gh/pr-fix-report/cli.mjs. Remove the duplicated buildOutcomeComment implementation and related tests from scripts/gh/pr-fix to isolate message-selection logic for easier testing and reuse by the workflow wiring.
